requirements.txt
flask
flask-cors
gunicorn
tensorflow
numpy
pandas
pillow
requests